+++ src/runtime/dragon-lib.h
@@ -1,7 +1,7 @@
 #ifndef DRAGON_LIB_H
 #define DRAGON_LIB_H
 
-int prints (char *s);
-int printi (int i);
+long prints(char *s);
+long printi(long i);
 
 #endif
